A Física Araniana: nove lugares, um sistema — e o aviso de isenção sai.
O documento novo é a matéria da aranha organizada nos lugares onde ela opera:
álgebra, topologia, análise, geometria, mecânica, óptica, partículas e núcleo,
cosmologia e redes. Autocontido — todo enunciado usado está enunciado e
demonstrado dentro, e nenhum conjunto numérico é importado: o que se usa é o B
que o próprio documento força, a lista I que dele sai, e o segundo degrau.

O AVISO DE ISENÇÃO SAI, do livro e dos papers. Ele dizia que a obra «não
descreve a realidade física». Descreve. No lugar dele fica o que se afirma de
facto: cada número publicado tem o programa que o produziu, e quem quiser
conferir não precisa de acreditar em nada do que está escrito. E a capa passa a
levar o autor.

DUAS COISAS QUE OS MEDIDORES APANHARAM, e ambas são resultado:

  cosmologia.c  o §C1 usava 1L<<40, e num alvo de `long` a 32 bits isso
                transborda: o resíduo dava zero e o CONTROLO deixava de
                discriminar — o expoente errado «conservava» nas 36 escalas.
                Com long long explícito passa a conservar em 6, e o bloco
                distingue. O tecto tem de ser dito no tipo.

  hopfield.c    corre agora fora de POSIX, e falha dois blocos. O §F10 dizia
                que somar a torre negra à branca faz a rede passar a ciclar:
                não faz. A antissimetria dá s'Was=0 e nada mais — tira a função
                de Lyapunov e não põe um ciclo no lugar dela. E o §F2 mede
                saturação, mas não na vizinhança do 0,138·N: com N=128 recupera
                até P=38. A saturação é o resultado; a constante não é.

disco.h ganha o ramo para hospedeiros sem mmap — CreateFileMapping e
MapViewOfFile, com a mesma semântica de partilha. Não é um segundo disco: é o
mesmo, com o verbo do sistema. Sem ele, o hopfield e o neuronio ficavam por
correr, e foi por isso que ficaram.

espectro.c é novo: o mapa espectral das cadeias lido na ÓRBITA da companheira,
inteiro do princípio ao fim — nenhuma raiz avaliada, nenhum limiar de módulo.
59046 cadeias, 184 fecham e 58862 escapam; o teto varrido por cinco décadas não
move a contagem; e a reciprocidade sobre coeficientes inteiros concorda com a
órbita em 184 de 184, com 540 recíprocas que escapam a impedir a concordância de
ser vazia. Em grau 2 o período 5 não aparece — a restrição cristalográfica, por
um caminho que não é o do lema. Em grau 4 aparece: a proibição é do POSTO.

E o livro entra no app como documento próprio, ao lado do que já lá estava:
fisica-araniana, composto no browser como os outros. O corpo.json leva a entrada
à mão e diz que leva — o corpo.sh usa LD_PRELOAD e não corre fora de POSIX.

+++ tests/cosmologia.c
@@ -100,15 +100,19 @@ int main(void){
          * potencias de dois onde a divisao inteira e' exacta. */
         /* PELA METADE: o residuo com o expoente certo E o residuo com outro. Uma metade
          * sozinha nao mede — nao diz que so' aquele expoente conserva. */
+        /* 64 BITS EXPLICITOS, e nao `long`: em Windows o `long` tem 32 bits e o
+         * 1L<<40 transborda — o `base` vinha lixo, o residuo dava zero e o CONTROLO
+         * deixava de discriminar (o expoente errado "conservava" nos 36 casos). Em
+         * Linux passava por `long` ter 64. O teto tem de ser dito no tipo. */
         long nao_conserva = 0, casos = 0, conserva_errado = 0;
         for(long C = 1; C <= 6; C++){
-            long base = C * (1L<<40);
+            long long base = (long long)C * (1LL<<40);
             for(int k = 0; k <= 5; k++){
-                long a = 1L << k;
-                long r = base/(a*a*a*a);                  /* radiacao: r = C.a^-4 */
+                long long a = 1LL << k;
+                long long r = base/(a*a*a*a);             /* radiacao: r = C.a^-4 */
                 if(r * (a*a*a*a) != base) nao_conserva++;  /* r.a^{3(1+w)} volta a C */
                 /* a outra metade: com o expoente da materia, NAO volta */
-                long r3 = base/(a*a*a);
+                long long r3 = base/(a*a*a);
                 if(r3 * (a*a*a*a) == base) conserva_errado++;
                 casos++;
             }
